pyproject.toml
[project]
name = "extraction"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiohttp>=3.13.3",
    "asgiref>=3.11.1",
    "asyncio>=4.0.0",
    "channels>=4.3.2",
    "channels-redis>=4.3.0",
    "cors>=1.0.1",
    "daphne>=4.2.1",
    "django>=6.0.2",
    "django-cors-headers>=4.9.0",
    "django-extensions>=4.1",
    "django-redis>=6.0.0",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "dotenv>=0.9.9",
    "google-genai>=1.61.0",
    "psycopg2-binary>=2.9.11",
    "psycopg[binary]>=3.3.2",
    "pyjwt>=2.11.0",
    "python-dotenv>=1.2.1",
    "pytz>=2025.2",
    "redis>=7.1.0",
    "requests>=2.32.5",
    "sqlparse>=0.5.5",
]
